requirements.txt
Flask==1.1.2
feedparser==6.0.10
python-dateutil==2.8.2
APScheduler==3.11.0